requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
backoff==2.2.1
bcrypt==4.3.0
billiard==4.2.1
boto3==1.37.2
botocore==1.37.2
build==1.2.2.post1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0
Django==4.2.19
django-storages==1.14.5
djangorestframework==3.15.2
dotenv==0.9.9
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.9
filelock==3.16.1
flatbuffers==25.2.10
flower==2.0.1
frozenlist==1.5.0
fsspec==2025.2.0
google-auth==2.38.0
googleapis-common-protos==1.68.0
graphlib_backport==1.1.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.1
humanfriendly==10.0
humanize==4.10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jiter==0.8.2
jmespath==1.0.1
kombu==5.4.2
kubernetes==32.0.1
lxml==5.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.16.3
openai==0.28.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pdf2image==1.17.0
pillow==10.4.0
posthog==3.17.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pysqlite3-binary==0.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.11.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.44.0
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.20.3
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
uvicorn==0.33.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.20.2
